DATACOUCH-524 - ability to store Optional<T> fields as field:null if the field is Optional.empty()
MOTIVATION
==========
This was a customer request.
Store Optional<T> fields as field:null if the field is Optional.empty().
If the field is null, do not emit the field at all (normal handling)

CHANGES
=======
Use the Jdk8Module of the ObjectMapper to handle Optional<T>.
In MappingCouchbaseConverter, recognize Optional<T> fields when reading and writing